Fix Sonar issue
