+++ src/main/java/com/powsybl/openloadflow/network/impl/AbstractFictitiousLfBranch.java
@@ -11,77 +11,15 @@ import com.powsybl.openloadflow.network.LfNetwork;
 import com.powsybl.openloadflow.network.PiModel;
 import com.powsybl.openloadflow.util.Evaluable;
 
-import java.util.Objects;
-
 import static com.powsybl.openloadflow.util.EvaluableConstants.NAN;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public abstract class AbstractFictitiousLfBranch extends AbstractLfBranch {
-
-    protected Evaluable p1 = NAN;
-
-    protected Evaluable q1 = NAN;
-
-    protected Evaluable i1 = NAN;
-
-    protected Evaluable p2 = NAN; // used for zero impedance branch only.
-
-    protected Evaluable q2 = NAN; // used for zero impedance branch only.
+public abstract class AbstractFictitiousLfBranch extends AbstractImpedantLfBranch {
 
     protected AbstractFictitiousLfBranch(LfNetwork network, LfBus bus1, LfBus bus2, PiModel piModel) {
         super(network, bus1, bus2, piModel);
-    }
-
-    @Override
-    public void setP1(Evaluable p1) {
-        this.p1 = Objects.requireNonNull(p1);
-    }
-
-    @Override
-    public Evaluable getP1() {
-        return p1;
-    }
-
-    @Override
-    public void setP2(Evaluable p2) {
-        this.p2 = Objects.requireNonNull(p2);
-    }
-
-    @Override
-    public Evaluable getP2() {
-        return p2;
-    }
-
-    @Override
-    public void setQ1(Evaluable q1) {
-        this.q1 = Objects.requireNonNull(q1);
-    }
-
-    @Override
-    public Evaluable getQ1() {
-        return q1;
-    }
-
-    @Override
-    public void setQ2(Evaluable q2) {
-        this.q2 = Objects.requireNonNull(q2);
-    }
-
-    @Override
-    public Evaluable getQ2() {
-        return q2;
-    }
-
-    @Override
-    public void setI1(Evaluable i1) {
-        this.i1 = Objects.requireNonNull(i1);
-    }
-
-    @Override
-    public Evaluable getI1() {
-        return i1;
     }
 
     @Override
@@ -93,5 +31,4 @@ public abstract class AbstractFictitiousLfBranch extends AbstractLfBranch {
     public Evaluable getI2() {
         return NAN;
     }
-
 }
